Implemented a generic linked list in C just to see if I could
Basically it just takes a void* so basically an any from typescript but in C,
Then it just depends on you manually converting that void pointer into the data
type that it needs. So you still gotta know which data type you are dealing with
But the same code for linked list can be used with different data types.