Went back and looked at the "official" implementation of BST deletion and adopted it to C
It was interesting the way they did all of it inside a recursive function. But
I found out that the issues that were making my old code cumbersome and unwieldy
still remained in this new solution so there was no escaping it. When you try to
choose whether to find the predecessor or successor based on the height of left
and right, there is added complexity that is unavoidable. Also there was an
interesting edge case where the value of a node and the value of its right child
would be the same, breaking the rules of BST. And so I had to add special logic
to handle this edge case